+++ src/main/java/day9/Task2/Rectangle.java
@@ -1,10 +1,10 @@
 package day9.Task2;
 
 public class Rectangle extends  Figure{
-    private int width;
-    private  int height;
+    private double width;
+    private  double height;
 
-    public Rectangle(int width, int height, String color) {
+    public Rectangle(double width, double height, String color) {
         super(color);
         this.height = height;
         this.width = width;
